Separate decode / decode image interface. Return error if decoded tile is empty.
